backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
dnspython==2.6.1
dkimpy==1.1.6
pyspf==2.0.14
tldextract==5.1.2
beautifulsoup4==4.12.3
langchain==0.2.15
langchain-ollama==0.1.1
httpx==0.27.2
python-multipart==0.0.9
gradio==4.44.0
requests==2.32.3
